+++ reactive-banana/src/Reactive/Banana/Internal/Combinators.hs
@@ -219,13 +219,15 @@ executeE e = do
     p <- liftBuildFun Prim.buildLaterReadNow $ executeP =<< runCached e
     return $ fromPure p
 
-switchE :: Event (Event a) -> Moment (Event a)
-switchE e = ask >>= \r -> cacheAndSchedule $ do
+switchE :: Event a -> Event (Event a) -> Moment (Event a)
+switchE e0 e = ask >>= \r -> cacheAndSchedule $ do
+    p0 <- runCached e0
     p1 <- runCached e
     liftBuild $ do
         p2 <- Prim.mapP (runReaderT . runCached) p1
+
         p3 <- Prim.executeP p2 r
-        Prim.switchP p3
+        Prim.switchP p0 p3
 
 switchB :: Behavior a -> Event (Behavior a) -> Moment (Behavior a)
 switchB b e = ask >>= \r -> cacheAndSchedule $ do
@@ -239,7 +241,8 @@ switchB b e = ask >>= \r -> cacheAndSchedule $ do
         -- TODO: switch away the initial behavior
         let c1 = p0                              -- initial behavior changes
         c2 <- Prim.mapP (const ()) p3            -- or switch happens
-        c3 <- Prim.switchP =<< Prim.mapP snd p3  -- or current behavior changes
+        never <- Prim.neverP
+        c3 <- Prim.switchP never =<< Prim.mapP snd p3  -- or current behavior changes
         pr <- merge c1 =<< merge c2 c3
         return (lr, pr)