+++ reactive-banana/src/Reactive/Banana/Prim/Combinators.hs
@@ -8,7 +8,7 @@ import Control.Monad
 import Control.Monad.IO.Class
 
 import Reactive.Banana.Prim.Plumbing
-    ( neverP, newPulse, newLatch, cachedLatch
+    ( newPulse, newLatch, cachedLatch
     , dependOn, keepAlive, changeParent
     , getValueL
     , readPulseP, readLatchP, readLatchFutureP, liftBuildP,
@@ -127,10 +127,9 @@ executeP p1 b = do
     eval (Just x) = Just <$> liftBuildP (x b)
     eval Nothing  = return Nothing
 
-switchP :: Pulse (Pulse a) -> Build (Pulse a)
-switchP pp = mdo
-    never <- neverP
-    lp    <- stepperL never pp
+switchP :: Pulse a -> Pulse (Pulse a) -> Build (Pulse a)
+switchP p pp = mdo
+    lp <- stepperL p pp
     let
         -- switch to a new parent
         switch = do
@@ -145,5 +144,6 @@ switchP pp = mdo
     p1 <- newPulse "switchP_in" switch :: Build (Pulse ())
     p1 `dependOn` pp
     p2 <- newPulse "switchP_out" eval
+    p2 `dependOn` p
     p2 `keepAlive` p1
     return p2