+++ reactive-banana/tests/Plumbing.hs
@@ -83,10 +83,10 @@ valueB ~(B x y) = M (X.valueB x) (Y.valueB y)
 observeE :: Event (Moment a) -> Event a
 observeE (E x y) = E (X.observeE $ fmap fstM x) (Y.observeE $ fmap sndM y)
 
-switchE :: Event (Event a) -> Moment (Event a)
-switchE (E x y) = M
-    (fmap ex $ X.switchE $ fmap fstE x)
-    (fmap ey $ Y.switchE $ fmap sndE y)
+switchE :: Event a -> Event (Event a) -> Moment (Event a)
+switchE (E x0 y0) (E x y) = M
+    (fmap ex $ X.switchE x0 $ fstE <$> x)
+    (fmap ey $ Y.switchE y0 $ sndE <$> y)
 
 switchB :: Behavior a -> Event (Behavior a) -> Moment (Behavior a)
 switchB (B x y) (E xe ye) = M